+++ ft-sdk/src/main/java/com/ft/sdk/FTRUMInnerManager.java
@@ -134,8 +134,11 @@ public class FTRUMInnerManager {
         boolean longTimeSession = now - lastSessionTime > SESSION_EXPIRE_TIME;
         if (longTimeSession || longResting) {
             lastSessionTime = now;
+            lastActionTime = now;
 
             sessionId = Utils.randomUUID();
+            LogUtils.d(TAG, "New SessionId:" + sessionId);
+
             checkSessionKeep(sessionId, sampleRate);
 
             if (checkRefreshView) {
@@ -155,7 +158,7 @@ public class FTRUMInnerManager {
                     FTMonitorManager.get().addMonitor(activeView.getId());
                     FTMonitorManager.get().attachMonitorData(activeView);
                     initView(activeView);
-                    LogUtils.d(TAG, "New sessionId:" + activeView.getSessionId() + ",viewId:" + activeView.getId());
+                    LogUtils.d(TAG, "checkRefreshView sessionId:" + activeView.getSessionId() + ",viewId:" + activeView.getId());
                 }
             }
 
@@ -394,6 +397,7 @@ public class FTRUMInnerManager {
         FTMonitorManager.get().addMonitor(activeView.getId());
         FTMonitorManager.get().attachMonitorData(activeView);
         initView(activeView);
+        lastActionTime = activeView.getStartTime();
 
     }